melhorado função pot.c

+++ src/functions.h
@@ -5,4 +5,4 @@ int sub(int a, int b);
 int mul(int a, int b);
 double divi(double a, double b);
 double raiz(double a);
-double pot(int a, int b);
+double pot(double a, int b);
